restucture NavBar components
